requirement.txt
pandas
sqlalchemy
pymysql
boto3
python-dotenv
flask
pyspark
